feat(leave): half-days, comp-off, probation and carry-forward — C4 complete
Completes the leave policy engine. Half-days and comp-off are the last two R2
gate scenarios, and both needed things outside the leave module, which is why
they were not in the first pass.

FRACTIONAL DAYS. A half-day of unpaid leave is half a day of loss of pay, so
leave_requests.days, payslips.paid_days and payslips.lop_days stop being
integers and the whole LOP chain in payroll carries Decimal. payslips
.working_days deliberately does NOT change: it counts days in a calendar and is
always whole, and a fractional column there would invite somebody to write half
a working day into it.

Two things broke on the way and both were mine. The audit payload passed a
Decimal to plain json.dumps, which JSONB serialisation cannot take - str() it,
exactly as `tds` already was. And pydantic renders Decimal as a STRING, which
silently turned days: 2 into days: "2" for every existing caller. Money crosses
the wire as a string because rupees-and-paise cannot round-trip a float; days
cannot have that problem, since half a day is the smallest unit anybody grants
and 0.5 is exact in binary. So days are numbers and money stays strings, and
the reason is written down where the next person will look.

A half-day is refused over a range. Half of a five-day absence is not a thing
anybody means, and accepting it would bill 2.5 days for a week away.

COMP-OFF is the three-module join the plan calls the product: an attendance
fact, measured against a calendar rule, becomes a leave credit. No one module
can decide it - attendance does not know the day was a holiday, the calendar
does not know anybody worked it, leave knows neither. It is EARNED rather than
granted, so a policy cannot allocate it and the route refuses to let one try;
unapproved work earns nothing, which is what approval is for; and a work fact
is the unit, so re-reading cannot credit the same Sunday twice. Four hours
earns half a day, seven earns a full one, and a two-hour call-out earns nothing
because that is overtime, not a day owed back.

EXIT DATES. employees.status could say "exited" and nothing recorded WHEN, so
exit proration was implemented and unreachable. The migration back-fills
nothing: inventing a date would put a fabricated figure into an F&F
calculation, and a leaver with no recorded date prorates exactly as before.

PROBATION and CARRY-FORWARD are wired rather than stored. Probation returns
zero with the reason intact - not owed yet, as opposed to spent. Carry-forward
is derived from last year's numbers and capped, for the same reason balances
are derived: a carried-forward counter is one more thing to drift out of step
with the requests it summarises.

Two levers are recorded and deliberately NOT enforced, and saying so is the
point. carry_forward_expires_months needs a credit ledger with dates to know
WHEN a day was carried; expiring one that should not have been is a day
somebody loses silently, so carrying it too long is the safer half of a wrong
answer. `encashable` is read by F&F in Phase 6.

499 tests, 13 new. The 486 that existed pass unchanged.

+++ apps/api/app/modules/payroll/models.py
@@ -181,8 +181,11 @@ class Payslip(TenantBase):
     employee_name: Mapped[str] = mapped_column(String(200))
 
     working_days: Mapped[int] = mapped_column(Integer)
-    paid_days: Mapped[int] = mapped_column(Integer)
-    lop_days: Mapped[int] = mapped_column(Integer, default=0)
+    #: Fractional, because half-day leave exists. `working_days` above stays
+    #: whole: it counts days in a calendar, and half a working day is not a
+    #: thing a calendar can contain.
+    paid_days: Mapped[Decimal] = mapped_column(Numeric(5, 1))
+    lop_days: Mapped[Decimal] = mapped_column(Numeric(5, 1), default=0)
     #: True once HR has typed an LOP figure by hand. Recomputing a draft must
     #: not silently discard it — an exit, an unpaid sabbatical, or a correction
     #: is knowledge the system doesn't have.
